Dostęp do zasobu

+++ RestaurantAPI/Services/RestaurantService.cs
@@ -1,18 +1,21 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
+using RestaurantAPI.Authorization;
 using RestaurantAPI.Entities;
 using RestaurantAPI.Exceptions;
 using RestaurantAPI.Models;
+using System.Security.Claims;
 
 namespace RestaurantAPI.Services
 {
     public interface IRestaurantService
     {
-        int Create(CreateRestaurantDto dto);
+        int Create(CreateRestaurantDto dto, int userId);
         IEnumerable<RestaurantDto> GetAll();
         RestaurantDto GetById(int id);
-        public void Delete(int id);
-        public void Update(UpdateRestaurantDto dto, int id);
+        public void Delete(int id, ClaimsPrincipal user);
+        public void Update(UpdateRestaurantDto dto, int id, ClaimsPrincipal user);
     }
 
     public class RestaurantService : IRestaurantService
@@ -20,12 +23,17 @@ namespace RestaurantAPI.Services
         private readonly RestaurantDbContext dbContext;
         private readonly IMapper mapper;
         private readonly ILogger<RestaurantService> logger;
+        private readonly IAuthorizationService authorizationService;
 
-        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger)
+        public RestaurantService(RestaurantDbContext dbContext, 
+            IMapper mapper, 
+            ILogger<RestaurantService> logger,
+            IAuthorizationService authorizationService)
         {
             this.dbContext = dbContext;
             this.mapper = mapper;
             this.logger = logger;
+            this.authorizationService = authorizationService;
         }
 
         public RestaurantDto GetById(int id)
@@ -58,16 +66,17 @@ namespace RestaurantAPI.Services
             return restaurantDtos;
         }
 
-        public int Create(CreateRestaurantDto dto)
+        public int Create(CreateRestaurantDto dto, int userId)
         {
             var restaurant = mapper.Map<Restaurant>(dto);
+            restaurant.CreatedById = userId;
             dbContext.Restaurants.Add(restaurant);
             dbContext.SaveChanges();
 
             return restaurant.Id;
         }
 
-        public void Delete(int id)
+        public void Delete(int id, ClaimsPrincipal user)
         {
             logger.LogError($"Restaurant with id: {id} Delete action invoked");
 
@@ -78,12 +87,20 @@ namespace RestaurantAPI.Services
             if (restaurant is null)
                 throw new NotFoundException("Restaurant not found");
 
+            var authorizationResult = authorizationService.AuthorizeAsync(user,
+                restaurant,
+                 new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             dbContext.Restaurants.Remove(restaurant);
             dbContext.SaveChanges();
         }
 
-        public void Update(UpdateRestaurantDto dto, int id)
+        public void Update(UpdateRestaurantDto dto, int id, ClaimsPrincipal user)
         {
             var restaurant = dbContext
                 .Restaurants
@@ -91,6 +108,15 @@ namespace RestaurantAPI.Services
 
             if (restaurant is null)
                 throw new NotFoundException("Restaurant not found");
+
+            var authorizationResult = authorizationService.AuthorizeAsync(user,
+                restaurant,
+                 new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             restaurant.Name = dto.Name;
             restaurant.Description = dto.Description;
